chore(suite): move coinjoin setup components to wallet views
