+++ packages/suite/src/views/wallet/coinjoin/components/DetailRow.tsx
